ci(release): move cut-tag scripts under .github/scripts/release
Keep release automation distinct from other github scripts. Tell operators to merge (not cherry-pick) when hotfix sync fails.
